Change CalculateSystemIntakeLCIDStatus() into a method on SystemIntake
