+++ pkg/graph/schema.resolvers.go
@@ -2974,8 +2974,8 @@ func (r *systemIntakeResolver) StatusAdmin(ctx context.Context, obj *models.Syst
 }
 
 // LcidStatus is the resolver for the lcidStatus field.
-func (r *systemIntakeResolver) LcidStatus(ctx context.Context, obj *models.SystemIntake) (*model.SystemIntakeLCIDStatus, error) {
-	return resolvers.CalculateSystemIntakeLCIDStatus(obj, time.Now()), nil
+func (r *systemIntakeResolver) LcidStatus(ctx context.Context, obj *models.SystemIntake) (*models.SystemIntakeLCIDStatus, error) {
+	return obj.LCIDStatus(time.Now()), nil
 }
 
 // DocumentType is the resolver for the documentType field.
