+++ pkg/models/system_intake_test.go
@@ -1,4 +1,4 @@
-package resolvers
+package models
 
 import (
 	"testing"
@@ -6,47 +6,44 @@ import (
 
 	"github.com/guregu/null"
 	"github.com/stretchr/testify/assert"
-
-	"github.com/cmsgov/easi-app/pkg/graph/model"
-	"github.com/cmsgov/easi-app/pkg/models"
 )
 
-func TestCalculateSystemIntakeLCIDStatus(t *testing.T) {
+func TestLCIDStatus(t *testing.T) {
 	mockCurrentTime := time.Unix(0, 0)
 	yesterday := mockCurrentTime.Add(time.Hour * -24)
 	tomorrow := mockCurrentTime.Add(time.Hour * 24)
 
 	t.Run("LCID status is nil if no LCID exists", func(t *testing.T) {
-		intake := &models.SystemIntake{
+		intake := &SystemIntake{
 			LifecycleID: null.StringFromPtr(nil),
 		}
 
-		lcidStatus := CalculateSystemIntakeLCIDStatus(intake, mockCurrentTime)
+		lcidStatus := intake.LCIDStatus(mockCurrentTime)
 
 		assert.Nil(t, lcidStatus)
 	})
 
 	t.Run("LCID status is ISSUED if LCID exists (but hasn't been retired) and expiration date has *not* been reached", func(t *testing.T) {
-		intake := &models.SystemIntake{
+		intake := &SystemIntake{
 			LifecycleID:        null.StringFrom("123456"),
 			LifecycleExpiresAt: &tomorrow,
 		}
 
-		lcidStatus := CalculateSystemIntakeLCIDStatus(intake, mockCurrentTime)
+		lcidStatus := intake.LCIDStatus(mockCurrentTime)
 
 		assert.NotNil(t, lcidStatus)
-		assert.EqualValues(t, model.SystemIntakeLCIDStatusIssued, *lcidStatus)
+		assert.EqualValues(t, SystemIntakeLCIDStatusIssued, *lcidStatus)
 	})
 
 	t.Run("LCID status is EXPIRED if LCID exists (but hasn't been retired) and expiration date *has* been reached", func(t *testing.T) {
-		intake := &models.SystemIntake{
+		intake := &SystemIntake{
 			LifecycleID:        null.StringFrom("234567"),
 			LifecycleExpiresAt: &yesterday,
 		}
 
-		lcidStatus := CalculateSystemIntakeLCIDStatus(intake, mockCurrentTime)
+		lcidStatus := intake.LCIDStatus(mockCurrentTime)
 
 		assert.NotNil(t, lcidStatus)
-		assert.EqualValues(t, model.SystemIntakeLCIDStatusExpired, *lcidStatus)
+		assert.EqualValues(t, SystemIntakeLCIDStatusExpired, *lcidStatus)
 	})
 }
